# 7 семестр/Web/LR2/openapi.yaml
...
openapi: 3.0.0
info:
  title: Books
  description: API для книжной информационной системы
  version: 1.0.0
paths:
...
  /bookshelf/{BookshelfId}:
    get:
      tags:
        - Bookshelf
      security:
        - Bearer: []
      summary: Избранное пользователя 
      parameters:
        - name: BookshelfId
          in: path       
          required: true  
          schema:
            type: integer 
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookshelfDTO'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized 
        '404':
          description: Not Found
  /bookshelf/{BookshelfId}/{BookId}:
    post:
      tags:
        - Bookshelf
      security:
        - Bearer: []
      summary: Добавление книги 
      parameters:
        - name: BookshelfId
          in: path       
          required: true  
          schema:
            type: integer 
        - name: BookId
          in: path       
          required: true  
          schema:
            type: integer 
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookshelfDTO'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized 
        '404':
          description: Not Found
    patch:
      tags:
        - Bookshelf
      security:
        - Bearer: []
      summary: Чтение книги 
      parameters:
        - name: BookshelfId
          in: path       
          required: true  
          schema:
            type: integer 
        - name: BookId
          in: path       
          required: true  
          schema:
            type: integer 
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookshelfDTO'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized 
        '404':
          description: Not Found      
    delete:
      tags:
        - Bookshelf
      security:
        - Bearer: []
      summary: Удаление книги 
      parameters:
        - name: BookshelfId
          in: path       
          required: true  
          schema:
            type: integer 
        - name: BookId
          in: path       
          required: true  
          schema:
            type: integer 
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookshelfDTO'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized 
        '404':
          description: Not Found
...
components:
  schemas:
    Author:
      type: object
      properties:
        Id:
          type: integer
          format: int32
        Name:
          type: string
        YearBirth:
          type: integer
          format: int32
        YearDeath:
          type: integer
          format: int32
        Country:
          type: string
        Genre:
          type: string
      required:
        - Name
        - YearBirth
        - Country
        - Genre
    Book:
      type: object
      properties:
        Id:
          type: integer
          format: int32
        Name:
          type: string
        Genre:
          type: string
        Year:
          type: integer
          format: int32
        Language:
          type: string
        Rating:
          type: number
          format: double
      required:
        - Name
        - Genre
        - Year
        - Language
        - Rating
...
    Bookshelf:
      type: object
      properties:
        Id:
          type: integer
          format: int32
        IdUser:
          type: integer
          format: int32
        Number:
          type: integer
          format: int32
        Rating:
          type: number
          format: double
      required:
        - IdUser
        - Number
        - Rating
...
    BookshelfDTO:
      type: object
      properties:
        Bookshelf:
          $ref: '#/components/schemas/Bookshelf'
        Books:
          type: array
          items:
            $ref: '#/components/schemas/BookOnBookshelf'
    BookOnBookshelf:
      type: object
      properties:
        Book:
          $ref: '#/components/schemas/Book'
        IsRead:
          type: boolean
        Authors:
          type: array
          items:
            $ref: '#/components/schemas/Author'
...
  securitySchemes:
    Bearer:
      type: apiKey
      name: Authorization
      in: header